requirements.txt
Flask
Flask-SocketIO
google-generative-ai
python-dotenv
